Create SafeTransactionService, Move get safe info to SafeTransactionService
