+++ Multisig/Data/Services/SafeRelayService.swift
@@ -16,44 +16,4 @@ class SafeRelayService {
         self.logger = logger
         httpClient = JSONHTTPClient(url: url, logger: logger)
     }
-
-    func safeInfo(at address: String) throws -> SafeStatusRequest.Response {
-        return try httpClient.execute(request: SafeStatusRequest(address: address))
-    }
-}
-
-struct SafeStatusRequest: JSONRequest {
-    let address: String
-
-    var httpMethod: String { return "GET" }
-    var urlPath: String { return "/api/v1/safes/\(address)/" }
-
-    typealias ResponseType = Response
-
-    init(address: String) {
-        self.address = address
-    }
-
-    struct Response: Decodable {
-        let address: String
-        let masterCopy: String
-        let nonce: Int
-        let threshold: Int
-        let owners: [String]
-        let version: String
-
-        init(address: String,
-                    masterCopy: String,
-                    nonce: Int,
-                    threshold: Int,
-                    owners: [String],
-                    version: String) {
-            self.address = address
-            self.masterCopy = masterCopy
-            self.nonce = nonce
-            self.threshold = threshold
-            self.owners = owners
-            self.version = version
-        }
-    }
 }
